introduced DevConsole interface and renamed original client to DevConsoleV1
